Add files via upload
I constructed a one-of-a-kind bank database with a distinctive feature: it's intended for a bank with a single branch. This unorthodox technique was driven by my dedication to the CAP theorem's ideals. Because accomplishing all three components of the CAP theorem (Consistency, Availability, and Partition Tolerance) at the same time might be difficult, I've opted to emphasize Consistency and Availability above Partition Tolerance. Explore this project to explore how a single-branch bank database may achieve high consistency and availability without the complexity of partition tolerance, all while realizing that it is not always feasible to have all three at the same time owing to the CAP theorem.